Rename `sblaster.cpp|h` to `soundblaster.cpp|h`
