+++ src/hardware/audio/private/soundblaster.h
