+++ src/hardware/audio/soundblaster.h
